pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hokusai-flow"
version = "0.1.0"
description = "LangGraph-based AI development workflow automation with Claude Code integration"
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "shigenoko" }
]
keywords = ["hokusai", "langgraph", "workflow", "automation", "development", "human-in-the-loop", "claude-code"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
hokusai = "hokusai.cli_main:main"

[project.urls]
Homepage = "https://github.com/shigenoko/hokusai"
Documentation = "https://github.com/shigenoko/hokusai#readme"
Repository = "https://github.com/shigenoko/hokusai.git"
Issues = "https://github.com/shigenoko/hokusai/issues"

[tool.hatch.build.targets.wheel]
packages = ["hokusai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["hokusai"]
branch = true
omit = [
    "hokusai/__pycache__/*",
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 0
show_missing = true
